feat(permissions): Manage user permissions in a database Permissions can be broken down into two types and two implementations. Depending on the used implementation, these permission types will behave differently.
- Role Permissions
- User Permissions

Standard: This implementation will give the user-based permissions a higher priority and will override role-based permissions. Any permissions granted/rejected on the user will always take precendee over any role-based permissions assigned.

Strict: This implementation will reject a permission as soon as one rejected permission is found on either the user or any of the assigned roles. Granting a user a permission that is rejected on a role he is assigned to will not grant that user this permission.

Role-based permissions that define the same permission with different access rights will be rejected in case of any rejections on any role.

If a user is not assigned a permission, the user will inherit permissions from the role. If a user is assigned a permission of false or true, then the user's permission will override the role permission.